Refactoring code - small, highly optimised functions live in pypd/kernels/
